+++ CHANGES.md
@@ -14,6 +14,7 @@
 - Removed `publishing` configuration
 - Added `testSources` and `testResources` to simplify setting of these properties
 - Generated files are now stored in a separate directory in ./build/generated
+- Configurations can no longer have parents
 
 # 0.13 2020-06-23
 - Add utility functions for source file and classpath generation, see the `wemi.generation` package